refactor: extract route handlers from startHttp to improve code health
